2023/langchain/pyproject.toml
[tool.poetry]
name = "langchain"
version = "0.0.1"
description = ""
authors = ["Arjan <hi@arjancodes.com>"]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.7"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
langchain = "^0.0.187"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
]
